# t8ira-swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Codepeak Track 8 Child API Base",
        "description": "This is the REST API for Track 8 Child applications",
        "termsOfService": "",
        "contact": {
            "name": "Codepeak Technologies Pvt. Ltd.",
            "url": "https://codepeaktechnologies.com/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Development Server"
        },
        {
            "url": "https://cpdaily-api.onrender.com",
            "description": "Production Server"
        }
    ],
...
    "paths": {
        "/auth/ilogin": {
            "post": {
                "summary": "Login API for T8IRA.",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/credentials"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "user found"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "description": "the jwt token generated by the api",
                                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NDc3M2FhYmU3NGUyOGE1ODQyNzNhNGUiLCJ1c2VybmFtZSI6ImFkbWluQGNvZGVwZWFrLmNvbSIsInJvbGVzIjpbImFkbWluIl0sImlhdCI6MTY5NjkzNzE4MCwiZXhwIjoxNjk3NTQxOTgwfQ.RFqPIWNeCLnRtZw1ouIYIe-hVXRTYY41NYWASgU_1XA"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "description": "the name of the user",
                                                    "example": "codepeak"
                                                },
                                                "secchanged": {
                                                    "type": "boolean",
                                                    "description": "the bit to know if the user configured an own password",
                                                    "example": true
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "description": "the phone number of the user",
                                                    "example": 9999999999
                                                },
                                                "courseid": {
                                                    "type": "string",
                                                    "description": "Intern's course",
                                                    "example": "93429dskeur23dklsdhu394898"
                                                },
                                                "batchid": {
                                                    "type": "string",
                                                    "description": "Intern's batch",
                                                    "example": "93429dskeur23dklsdhu394898"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "description": "Intern's emailid",
                                                    "example": "name`@`mail.com"
                                                },
                                                "qualification": {
                                                    "type": "string",
                                                    "description": "Intern's educational qualification",
                                                    "example": "BE"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "user not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "summary": "General Logout API.",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Logged out successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/resetpassword": {
            "post": {
                "summary": "Reset Password API.",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "currentpassword": {
                                        "type": "string",
                                        "description": "The current password.",
                                        "example": "password"
                                    },
                                    "newpassword": {
                                        "type": "string",
                                        "description": "The new password.",
                                        "example": "password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Password updated successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Error in processing request"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "User not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/updatepassword": {
            "post": {
                "summary": "Update Password API.",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newpassword": {
                                        "type": "string",
                                        "description": "The new password.",
                                        "example": "password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Password updated successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Error in processing request"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Unauthorized"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "User not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "credentials": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "The username.",
                        "example": "emailid"
                    },
                    "password": {
                        "type": "string",
                        "description": "The password.",
                        "example": "password"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}